requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
sqlmodel==0.0.16
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
psycopg2-binary==2.9.9
email-validator==2.1.1
pandas==2.2.2
openpyxl==3.1.2
requests==2.32.3
openai==1.54.4
httpx==0.27.0
jinja2==3.1.4
python-jose[cryptography]==3.3.0
